Update and rename csv_parser.js to csvParser.js
#471 comments
